move viewModelState into a separate file
